+++ src/lang/compiler_wrapper.cc
@@ -14,14 +14,14 @@ using namespace argon::lang::parser;
 using namespace argon::lang::compiler;
 using namespace argon::object;
 
-Code *CompilerWrapper::Compile(const char *file_name, scanner2::Scanner *scanner) {
+Code *CompilerWrapper::Compile(const char *file_name, scanner::Scanner *scanner) {
     Parser parser(scanner, file_name);
     Compiler compiler;
     Code *code;
     File *ast;
 
     if ((ast = parser.Parse()) == nullptr) {
-        if (scanner->status != scanner2::ScannerStatus::GOOD)
+        if (scanner->status != scanner::ScannerStatus::GOOD)
             return (Code *) ErrorFormat(type_syntax_error_, "%s", scanner->GetStatusMessage());
         return nullptr;
     }
@@ -33,16 +33,16 @@ Code *CompilerWrapper::Compile(const char *file_name, scanner2::Scanner *scanner
 }
 
 Code *CompilerWrapper::Compile(const char *file_name, const char *code, unsigned long code_sz) {
-    scanner2::Scanner scanner(code, code_sz);
+    scanner::Scanner scanner(code, code_sz);
     return this->Compile(file_name, &scanner);
 }
 
 Code *CompilerWrapper::Compile(const char *file_name, FILE *fd) {
-    scanner2::Scanner scanner(fd, nullptr, nullptr);
+    scanner::Scanner scanner(fd, nullptr, nullptr);
     return this->Compile(file_name, &scanner);
 }
 
 Code *CompilerWrapper::Compile(const char *file_name, String *code) {
-    scanner2::Scanner scanner((const char *) code->buffer, code->len);
+    scanner::Scanner scanner((const char *) code->buffer, code->len);
     return this->Compile(file_name, &scanner);
 }
